Unify workflow step config resolution
Extract shared field and agent-default resolution paths so creation and dispatch fill sites use the same interpolation and defaulting behavior. No behavior change.

Refs ENG-766
