+++ .changeset/brisk-fields-unify.md
@@ -1,0 +1,5 @@
+---
+"@shipfox/api-workflows": patch
+---
+
+Unifies workflow step config field resolution and agent default completion across creation and dispatch fill sites.
