+++ libs/api/workflows/src/core/step-config/complete-step-dispatch-config.ts
@@ -1,22 +1,7 @@
-import {
-  InvalidAgentModelError,
-  UnsupportedModelProviderError,
-} from '@shipfox/api-agent/core/errors';
 import type {AgentDefaultsResolver} from '@shipfox/api-agent/core/resolve-agent-config';
-import type {AgentThinking} from '@shipfox/api-agent-dto';
-import {
-  freezePlannedRunCommandAtSite,
-  getWorkflowInterpolationFieldFailurePolicy,
-  resolveFieldAtSite,
-  type WorkflowInterpolationField,
-  WorkflowTemplateResolutionError,
-} from '@shipfox/expression';
-import type {Step, StepConfigDispatchPlan} from '#core/entities/step.js';
-import {
-  AgentConfigUnresolvableError,
-  InterpolationUnresolvableError,
-  type InterpolationUnresolvableField,
-} from '#core/errors.js';
+import type {Step} from '#core/entities/step.js';
+import {completeAgentConfig} from './agent.js';
+import {completeRunDispatchConfig} from './run.js';
 import type {WorkflowEvaluationContext} from './workflow-evaluation-context.js';
 
 export function completeStepDispatchConfig(params: {
@@ -29,34 +14,12 @@ export function completeStepDispatchConfig(params: {
   if (plan === null) return params.step.config;
 
   const config = {...params.step.config};
-  const env = {...readConfigEnv(config)};
-
-  if (plan.run !== undefined) {
-    const resolved = freezePlannedRunCommandAtSite({
-      field: plan.run,
-      site: params.context.site,
-      context: params.context.values,
-      failurePolicy: getWorkflowInterpolationFieldFailurePolicy('run'),
-      reservedNames: Object.keys(env),
-    });
-    config.run = resolved.command;
-    Object.assign(env, resolved.env);
-  }
-
-  if (plan.env !== undefined) {
-    for (const [key, field] of Object.entries(plan.env)) {
-      env[key] = completeField({
-        field: 'env.value',
-        errorField: 'env',
-        template: field,
-        context: params.context,
-        definitionId: params.definitionId,
-        envKey: key,
-      });
-    }
-  }
-
-  if (Object.keys(env).length > 0) config.env = env;
+  completeRunDispatchConfig({
+    config,
+    plan,
+    context: params.context,
+    definitionId: params.definitionId,
+  });
   completeAgentConfig({
     config,
     plan,
@@ -66,127 +29,4 @@ export function completeStepDispatchConfig(params: {
   });
 
   return config;
-}
-
-function completeAgentConfig(params: {
-  readonly config: Record<string, unknown>;
-  readonly plan: StepConfigDispatchPlan;
-  readonly context: WorkflowEvaluationContext;
-  readonly resolveAgentDefaults: AgentDefaultsResolver;
-  readonly definitionId: string;
-}): void {
-  const agent = params.plan.agent;
-  if (agent === undefined) return;
-
-  const prompt =
-    agent.prompt === undefined
-      ? readConfigString(params.config, 'prompt')
-      : completeField({
-          field: 'agent.prompt',
-          errorField: 'agent.prompt',
-          template: agent.prompt,
-          context: params.context,
-          definitionId: params.definitionId,
-        });
-  const model =
-    agent.model === undefined
-      ? readConfigString(params.config, 'model')
-      : completeField({
-          field: 'agent.model',
-          errorField: 'agent.model',
-          template: agent.model,
-          context: params.context,
-          definitionId: params.definitionId,
-        });
-  const provider =
-    agent.provider === undefined
-      ? readConfigString(params.config, 'provider')
-      : completeField({
-          field: 'agent.provider',
-          errorField: 'agent.provider',
-          template: agent.provider,
-          context: params.context,
-          definitionId: params.definitionId,
-        });
-
-  try {
-    const defaults = params.resolveAgentDefaults({
-      provider,
-      model,
-      thinking: agent.thinking ?? readConfigThinking(params.config),
-    });
-    params.config.provider = defaults.provider;
-    params.config.model = defaults.model;
-    params.config.thinking = defaults.thinking;
-    params.config.prompt = prompt;
-  } catch (error) {
-    if (error instanceof UnsupportedModelProviderError || error instanceof InvalidAgentModelError) {
-      throw new AgentConfigUnresolvableError(params.definitionId, {cause: error});
-    }
-    throw error;
-  }
-}
-
-function completeField(params: {
-  readonly field: WorkflowInterpolationField;
-  readonly errorField: InterpolationUnresolvableField;
-  readonly template: StepConfigDispatchPlan['run'];
-  readonly context: WorkflowEvaluationContext;
-  readonly definitionId: string;
-  readonly envKey?: string;
-}): string {
-  if (params.template === undefined) return '';
-  try {
-    const resolved = resolveFieldAtSite({
-      field: params.template,
-      site: params.context.site,
-      context: params.context.values,
-      failurePolicy: getWorkflowInterpolationFieldFailurePolicy(params.field),
-    });
-    if (resolved.kind === 'frozen') return resolved.value;
-    const source = resolved.field.segments.find((segment) => segment.kind === 'deferred')
-      ?.expression.source;
-    throw new InterpolationUnresolvableError(params.definitionId, {
-      field: params.errorField,
-      source: source ?? params.field,
-      ...(params.envKey === undefined ? {} : {envKey: params.envKey}),
-    });
-  } catch (error) {
-    if (error instanceof WorkflowTemplateResolutionError) {
-      throw new InterpolationUnresolvableError(params.definitionId, {
-        field: params.errorField,
-        source: error.source,
-        ...(params.envKey === undefined ? {} : {envKey: params.envKey}),
-        cause: error,
-      });
-    }
-    throw error;
-  }
-}
-
-function readConfigEnv(config: Record<string, unknown>): Record<string, string> {
-  const env = config.env;
-  if (env === null || typeof env !== 'object' || Array.isArray(env)) return {};
-  return Object.fromEntries(
-    Object.entries(env).flatMap(([key, value]) =>
-      typeof value === 'string' ? [[key, value]] : [],
-    ),
-  );
-}
-
-function readConfigString(config: Record<string, unknown>, key: string): string | undefined {
-  const value = config[key];
-  return typeof value === 'string' ? value : undefined;
-}
-
-function readConfigThinking(config: Record<string, unknown>): AgentThinking | undefined {
-  const value = config.thinking;
-  return value === 'off' ||
-    value === 'minimal' ||
-    value === 'low' ||
-    value === 'medium' ||
-    value === 'high' ||
-    value === 'xhigh'
-    ? value
-    : undefined;
 }
